Assignment 42 Labels completed!

+++ app/controllers/comments_controller.rb
@@ -4,29 +4,59 @@ class CommentsController < ApplicationController
   before_action :authorize_user, only: [:destroy]
 
   def create
-    @post = Post.find(params[:post_id])
-    comment = @post.comments.new(comment_params)
-    comment.user = current_user
-
-    if comment.save
-      flash[:notice] = "Comment saved successfuly."
-      redirect_to [@post.topic, @post]
+    if params[:post_id]
+      @post = Post.find(params[:post_id])
+      comment = @post.comments.new(comment_params)
     else
-      flash[:alert] = "Comment failed to save."
-      redirect_to [@post.topic, @post]
+      @topic = Topic.find(params[:topic_id])
+      comment = @topic.comments.new(comment_params)
+    end
+
+    comment.user = current_user
+    if @post
+      if comment.save
+        flash[:notice] = "Comment saved successfuly."
+        redirect_to [@post.topic, @post]
+      else
+        flash[:alert] = "Comment failed to save."
+        redirect_to [@post.topic, @post]
+      end
+    else
+      if comment.save
+        flash[:notice] = "Comment saved successfuly."
+        redirect_to [@topic]
+      else
+        flash[:alert] = "Comment failed to save."
+        redirect_to [@topic]
+      end
     end
   end
 
   def destroy
-    @post = Post.find(params[:post_id])
-    comment = @post.comments.find(params[:id])
-
-    if comment.destroy
-      flash[:notice] = "Comment was deleted successfully."
-      redirect_to [@post.topic, @post]
+    if params[:post_id]
+      @post = Post.find(params[:post_id])
+      comment = @post.comments.find(params[:id])
     else
-      flash[:alert] = "Comment couldn't be deleted. Try again."
-      redirect_to [@post.topic,@post]
+      @topic = Topic.find(params[:topic_id])
+      comment = @topic.comments.find(params[:id])
+    end
+
+    if @post
+      if comment.destroy
+        flash[:notice] = "Comment was deleted successfully."
+        redirect_to [@post.topic, @post]
+      else
+        flash[:alert] = "Comment couldn't be deleted. Try again."
+        redirect_to [@post.topic,@post]
+      end
+    else
+      if comment.destroy
+        flash[:notice] = "Comment was deleted successfully."
+        redirect_to [@topic]
+      else
+        flash[:alert] = "Comment couldn't be deleted. Try again."
+        redirect_to [@topic]
+      end
     end
   end
